+++ demo-app/src/main/java/cash/z/ecc/android/sdk/demoapp/fixture/WalletSnapshotFixture.kt
@@ -24,7 +24,6 @@ object WalletSnapshotFixture {
         processorInfo: CompactBlockProcessor.ProcessorInfo = CompactBlockProcessor.ProcessorInfo(
             null,
             null,
-            null,
             null
         ),
         orchardBalance: WalletBalance = ORCHARD_BALANCE,